[WIP] Implement Command pattern
