+++ app/src/main/kotlin/at/bitfire/davdroid/webdav/WebDavMountRepository.kt
@@ -33,7 +33,7 @@ class WebDavMountRepository @Inject constructor(
     private val mountDao = db.webDavMountDao()
     private val documentDao = db.webDavDocumentDao()
 
-    /** authority of our WebDAV document provider ([BaseDavDocumentsProvider]) */
+    /** authority of our WebDAV document provider ([DavDocumentsProviderWrapper]) */
     private val authority = context.getString(R.string.webdav_authority)
 
     /**
